pyproject.toml
[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[project]
name = "redelex"
version = "0.5.1"
description = "ReDeLEx: A Framework for Relational Deep Learning Exploration"
authors = [{ name = "Jakub Peleška", email = "jakub.peleska@fel.cvut.cz" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
requires-python = ">=3.12"
keywords = []
dependencies = [
    "relbench>=1.1.0",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.40",
    "inflect>=7.5.0",
    "tqdm>=4.67.1",
    "attrs>=25.3.0",
]

[dependency-groups]
cpu = [
    "torch==2.4.1",
    "torchvision==0.19.1",
    "torchaudio==2.4.1",
    "pytorch-frame>=0.2.5",
    "torch-geometric>=2.6.1",
    "sentence-transformers>=4.0.1",
]
cu124 = [
    "torch==2.4.1",
    "torchvision==0.19.1",
    "torchaudio==2.4.1",
    "pytorch-frame>=0.2.5",
    "torch-geometric>=2.6.1",
    "sentence-transformers>=4.0.1",
]

dev = ["ipykernel>=6.29.5", "matplotlib>=3.10.1", "pre-commit>=4.2.0"]
experiments = [
    "lightning>=2.5.1.post0",
    "mlflow>=2.22.1",
    "ray[data,serve,train,tune]>=2.46.0",
    "getml>=1.5.1",
]


[tool.uv]
default-groups = ["dev", "cpu", "experiments"]
conflicts = [[{ group = "cpu" }, { group = "cu124" }]]

[tool.uv.sources]
torch = [
    { index = "torch-cpu", group = "cpu" },
    { index = "torch-cu124", group = "cu124" },
]
torchvision = [
    { index = "torch-cpu", group = "cpu" },
    { index = "torch-cu124", group = "cu124" },
]
torchaudio = [
    { index = "torch-cpu", group = "cpu" },
    { index = "torch-cu124", group = "cu124" },
]


[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "torch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true


[project.urls]
Homepage = "https://github.com/jakubpeleska/ctu-relational-py"
Issues = "https://github.com/jakubpeleska/ctu-relational-py/issues"

[tool.ruff]
line-length = 92
target-version = "py310"

[tool.ruff.lint.pydocstyle]
convention = "google"
